HW-1/backend/llm/pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.72.6",
    "python-dotenv>=1.1.0",
    "fastapi",
    "uvicorn[standard]"
]
